refactor : float64 -> float32

+++ gym_footsteps_planning/envs/footsteps_planning_env.py
@@ -26,7 +26,7 @@ class FootstepsPlanningEnv(gymnasium.Env):
             "has_obstacle": False,
             "obstacle_max_radius": 0.25,  # [m]
             "obstacle_radius": None,  # [m]
-            "obstacle_position": np.array([0, 0], dtype=np.float64),  # [m,m]
+            "obstacle_position": np.array([0, 0], dtype=np.float32),  # [m,m]
             # Which foot is targeted (any, left or right)
             "foot": "any",
             # Foot geometry
@@ -55,17 +55,17 @@ class FootstepsPlanningEnv(gymnasium.Env):
 
         # Maximum speed in each dimension
         self.min_step = np.array(
-            [-self.options["max_dx_backward"], -self.options["max_dy"], -self.options["max_dtheta"]], dtype=np.float64
+            [-self.options["max_dx_backward"], -self.options["max_dy"], -self.options["max_dtheta"]], dtype=np.float32
         )
         self.max_step = np.array(
-            [self.options["max_dx_forward"], self.options["max_dy"], self.options["max_dtheta"]], dtype=np.float64
+            [self.options["max_dx_forward"], self.options["max_dy"], self.options["max_dtheta"]], dtype=np.float32
         )
 
         # Action space is target step size (dx, dy, dtheta)
         # To keep 0 as "not moving", we use maxStep instead of maxBackwardStep,
         # but the speed is clipped when stepping
         self.action_high = np.array(
-            [self.options["max_dx_forward"], self.options["max_dy"], self.options["max_dtheta"]], dtype=np.float64
+            [self.options["max_dx_forward"], self.options["max_dy"], self.options["max_dtheta"]], dtype=np.float32
         )
         self.action_low = -self.action_high
 
@@ -83,11 +83,11 @@ class FootstepsPlanningEnv(gymnasium.Env):
         # - the obstacle radius
         max_diag_env = np.sqrt(self.panjang**4 + self.lebar**4)
         self.state_low_goal = np.array(
-            [-max_diag_env, -max_diag_env, -1, -1, 0, -max_diag_env, -max_diag_env, 0], dtype=np.float64
+            [-max_diag_env, -max_diag_env, -1, -1, 0, -max_diag_env, -max_diag_env, 0], dtype=np.float32
         )
         self.state_high_goal = np.array(
             [max_diag_env, max_diag_env, 1, 1, 1, max_diag_env, max_diag_env, self.options["obstacle_max_radius"]],
-            dtype=np.float64,
+            dtype=np.float32,
         )
 
         self.observation_space = spaces.Box(self.state_low_goal, self.state_high_goal)
@@ -108,7 +108,7 @@ class FootstepsPlanningEnv(gymnasium.Env):
                 T_support_target[0, 0],  # cos(theta)
                 T_support_target[1, 0],  # sin(theta)
             ],
-            dtype=np.float64,
+            dtype=np.float32,
         )
 
         if self.options["has_obstacle"]:
@@ -138,10 +138,10 @@ class FootstepsPlanningEnv(gymnasium.Env):
                 is_target_foot,
             ]
             + ([*self.support_obstacle, self.obstacle_radius] if self.options["has_obstacle"] else [0, 0, 0]),
-            dtype=np.float64,
+            dtype=np.float32,
         )
 
-        state = np.array(state, dtype=np.float64)
+        state = np.array(state, dtype=np.float32)
 
         return state
 
@@ -157,7 +157,7 @@ class FootstepsPlanningEnv(gymnasium.Env):
                 self.options["max_dy"],
                 self.options["max_dtheta"],
             ],
-            dtype=np.float64,
+            dtype=np.float32,
         )
         clipped_step = step / factor
 
@@ -192,7 +192,7 @@ class FootstepsPlanningEnv(gymnasium.Env):
 
         distance = np.linalg.norm(state[:2])
         target_theta = self.target_foot_pose[2]
-        theta_error = np.arccos(np.clip(np.dot(state[2:4], np.array([1, 0])), -1.0, 1.0), dtype=np.float64)
+        theta_error = np.arccos(np.clip(np.dot(state[2:4], np.array([1, 0])), -1.0, 1.0), dtype=np.float32)
 
         is_desired_foot = state[4] == 1
 
@@ -207,7 +207,7 @@ class FootstepsPlanningEnv(gymnasium.Env):
                 for sy in [-1, 1]:
                     # One of the corners of feet is walking in the forbidden area, punishing this with negative reward
                     P_corner_foot = np.array(
-                        [sx * self.simulator.foot_length / 2, sy * self.simulator.foot_width / 2], dtype=np.float64
+                        [sx * self.simulator.foot_length / 2, sy * self.simulator.foot_width / 2], dtype=np.float32
                     )
 
                     if np.linalg.norm(P_corner_foot - support_obstacle) < obstacle_radius:
@@ -278,7 +278,7 @@ class FootstepsPlanningEnv(gymnasium.Env):
             if self.options["multi_goal"]:
                 self.target_foot_pose = self.np_random.uniform([-2, -2, -math.pi], [2, 2, math.pi])
             else:
-                self.target_foot_pose = np.array([0, 0, 0], dtype=np.float64)
+                self.target_foot_pose = np.array([0, 0, 0], dtype=np.float32)
 
         self.simulator.set_desired_goal(*self.target_foot_pose, self.target_support_foot)
